Add Python Representation data options.
This patch adds `eye0-pyrep`, `eye1-pyrep`, `gaze_x-pyrep`, and
`gaze_x-pyrep` to the options for 'data_names' in the YAML config.
This is only available if an XDF with the Python Representation is
added. Also, if no 'Primitive' data is foud, the data names with no
suffixed `-pyrep` will be set to the `diameter` option, which is the
data uncorrected for perspective.